+++ src/rules/no-unresolved.js
@@ -18,14 +18,17 @@ module.exports = {
     schema: [
       makeOptionsSchema({
         caseSensitive: { type: 'boolean', default: true },
+        caseSensitiveStrict: { type: 'boolean', default: false },
       }),
     ],
   },
 
   create(context) {
+    const options = context.options[0] || {};
+
     function checkSourceValue(source) {
-      const shouldCheckCase = !CASE_SENSITIVE_FS
-        && (!context.options[0] || context.options[0].caseSensitive !== false);
+      const caseSensitive = !CASE_SENSITIVE_FS && options.caseSensitive !== false;
+      const caseSensitiveStrict = !CASE_SENSITIVE_FS && options.caseSensitiveStrict;
 
       const resolvedPath = resolve(source.value, context);
 
@@ -34,9 +37,9 @@ module.exports = {
           source,
           `Unable to resolve path to module '${source.value}'.`
         );
-      } else if (shouldCheckCase) {
+      } else if (caseSensitive || caseSensitiveStrict) {
         const cacheSettings = ModuleCache.getSettings(context.settings);
-        if (!fileExistsWithCaseSync(resolvedPath, cacheSettings)) {
+        if (!fileExistsWithCaseSync(resolvedPath, cacheSettings, caseSensitiveStrict)) {
           context.report(
             source,
             `Casing of ${source.value} does not match the underlying filesystem.`
@@ -45,6 +48,6 @@ module.exports = {
       }
     }
 
-    return moduleVisitor(checkSourceValue, context.options[0]);
+    return moduleVisitor(checkSourceValue, options);
   },
 };