+++ tests/src/rules/no-unresolved.js
@@ -15,7 +15,7 @@ function runResolverTests(resolver) {
   function rest(specs) {
     specs.settings = Object.assign({},
       specs.settings,
-      { 'import/resolver': resolver },
+      { 'import/resolver': resolver, 'import/cache': { lifetime: 0 } },
     );
 
     return test(specs);
@@ -227,6 +227,10 @@ function runResolverTests(resolver) {
   });
 
   if (!CASE_SENSITIVE_FS) {
+    const relativePath = './tests/files/jsx/MyUnCoolComponent.jsx';
+    const cwd = process.cwd();
+    const mismatchedPath = path.join(cwd.toUpperCase(), relativePath).replace(/\\/g, '/');
+
     ruleTester.run('case sensitivity', rule, {
       valid: [
         rest({ // test with explicit flag
@@ -244,6 +248,29 @@ function runResolverTests(resolver) {
           code: 'import foo from "./jsx/MyUncoolComponent.jsx"',
           options: [{ caseSensitive: true }],
           errors: [`Casing of ./jsx/MyUncoolComponent.jsx does not match the underlying filesystem.`],
+        }),
+      ],
+    });
+
+    ruleTester.run('case sensitivity strict', rule, {
+      valid: [
+        // #1259 issue
+        rest({ // caseSensitiveStrict is disabled by default
+          code: `import foo from "${mismatchedPath}"`,
+        }),
+      ],
+
+      invalid: [
+        // #1259 issue
+        rest({ // test with enabled caseSensitiveStrict option
+          code: `import foo from "${mismatchedPath}"`,
+          options: [{ caseSensitiveStrict: true }],
+          errors: [`Casing of ${mismatchedPath} does not match the underlying filesystem.`],
+        }),
+        rest({ // test with enabled caseSensitiveStrict option and disabled caseSensitive
+          code: `import foo from "${mismatchedPath}"`,
+          options: [{ caseSensitiveStrict: true, caseSensitive: false }],
+          errors: [`Casing of ${mismatchedPath} does not match the underlying filesystem.`],
         }),
       ],
     });